Feature/workload

See merge request ElDiabolo/bachelorarbeit-backend!5

+++ database/migrations/2022_12_04_111711_create_contracts_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->foreignId('user_id')->constrained();
             $table->timestamp('from');
             $table->timestamp('to')->nullable();
-            $table->tinyInteger('hours_per_week');
+            $table->decimal('hours_per_week');
             $table->timestamps();
         });
     }
